Fixed oil duping bug
As i said easy fix also i removed some code related to lava. Coping and
pasting should be done with clipper in the other hand.
Also you need to change you color modifier so the oil is simi see threw.
Another thing is when in oil it resembles being in water. eg i can see
in oil the same as water.