Simple API version tracking
This isn't anything over the top or self maintaining. However, it should
cover the need, with a little manual input, which version of Each api UE
compatibility module was created with. Current this is missing and i've
noticed some issues with helping user when it comes to API version
errors/crashes.

This should later be expanded to do basic version checking and inform
the user that for best use to update or change mod versions. Though this
does rely on the mod using the API to call the init method.