Added to and improved Electricity Display helper
Raised Max unit measure up to Yotta. Though it will never be used beyond
peta its still nice to have the extra space if needed.

Change how getDisplay finds the correct unit value and should work for
future updates. Not methods, or names were changed so should be
backwards compatibly. Only issue that can arise is if someone is using
the Double value of MeasurementUnit enum.

Fixed getDisplay not working with neg numbers. It detects neg,
converting it to pos but modifies the final output with a '-' sign. Neg
sign can be removed using string.replace if needed.