Possible fix for wrong world be using on local client servers
On going issue that will need more toying with to solve completely. Issue is the client and lan both share resources between the server thread and client thread. This can cause the getWorld(int) method to return the wrong world for calls. When this happens isServer() isClient() checks can fail. As well other unknown issues could happen as a result of this problem.