Renamed render states, Added search function to debug menus, Improved debug menu display
More work is needed to fill out the debug system for JSON. This commit adds ability so filter display of information. It also fixes a lot of display problems by packing the GUI before display. Additionally, debug data is now inferface driven to allow improving rendering of data. Along with this JSON tab panel is now a seperate class. In the future all panels will be built this way to seperate out functionaliyt.