Created new file list method to solve issue #96
Still a lot of work left to do but this method can list the files inside a jar. This is a very good first step as the core problem has to do with the filesystem walk method. If this can be replaced then there is no need for URL to URI to find files.