Started work on implementing static 3D block models & rendering
This is being rushed into implementation mainly for AssemblyLine. However, i have a few ideas how to use this in ICBM-Classic to replace the old models with JSON versions. As well reduce FPS hits for TESR renders used by most tiles in both ICBM-Classic and MFFS. Which both use blocky 3D models that can with a little effort be converted over to static block renders (ISBR).