Created lambda driven computer data system
Replaces the short lived annotation system. All though a bit more of a pain to code it will be easier to work wtih. 

As methods will not need to be created in order to create spots for annotations to work. Instead an expression can be made to run logic or get data from the tile. As well this system allow creating global methods to add to every tile. So each tile will no longer need to supply some basic methods. Say for example getEnergy() and getMaxEnergy()